requirements.txt
# Trustworthy RAG System - Requirements
# Python 3.9+ required

# Core LangChain and OpenAI
langchain==0.2.16
langchain-openai==0.1.23
langchain-community>=0.2.16
openai>=1.35.0

# Google ADK (Agent Development Kit)
google-adk>=0.1.0
google-genai>=0.1.0
litellm>=1.40.0

# Vector Store and Embeddings
faiss-cpu==1.8.0
# For GPU support, use: faiss-gpu==1.8.0

# NLP and Text Processing
scikit-learn==1.3.2
numpy==1.24.3
tiktoken>=0.7.0

# Configuration & Validation
python-dotenv==1.0.0
dataclasses>=0.6
tqdm==4.66.1

# Pinecone for production vector database
pinecone-client==3.0.3

# Redis for caching (optional but recommended for production)
redis==5.0.1

# FastAPI for REST API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Metrics and monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Optional: For enhanced demo UI
# streamlit==1.29.0
# plotly==5.18.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0
mypy==1.7.1

